CI: Disable Psalm for PHP 8.4
Unfortunately Psalm 4 breaks on PHP 8.4 due to the implicit nullable parameters deprecation, but there's no newer Psalm release that still supports PHP 7.2. We either need to drop older PHP versions or do a complete overhaul of our testing to use different CI tools depending on the PHP version. This is out-of-scope right now.